Add an option to not clear after attach
I'm trying to make a "pick up where I left off but don't mess with my
scrollback buffers" kind of setup here. To this end, I'm using dtach
with script and scriptreplay to store session output and replay it to
my terminal.

The only hitch is that dtach normally clears the screen when it attaches
-- normally, this is nice behavior, but in this case it clears the stuff
that just played back.

So! This adds a -C option to dtach, which tells it _not_ to clear the
screen after attaching to an active session.

Might fix https://github.com/crigler/dtach/issues/6